Move unit tests for reconciliation
